refactor: clean up key utils
